Commit automatico

+++ application/models/Auth.php
@@ -280,7 +280,7 @@ class Auth extends CI_Model
                 'name' => static::$rememberTokenName,
                 //'value' => json_encode(['token_string' => $token_string, 'timeout' => time() + ($timeout*60)]),
                 'value' => json_encode(['token_string' => $token_string, 'timeout' => time() + ($timeout * 60)]),
-                'expire' => time() + (31 * 24 * 60 * 60),
+                'expire' => (int) (time() + (31 * 24 * 60 * 60)),
                 //'expire' => time() + ($timeout*60),
                 'domain' => '.' . $_SERVER['HTTP_HOST'],
                 'path' => '/'
